Port project to .NET Core
