--- .travis.yml
@@ -1,7 +1,0 @@
-language: csharp
-solution: X.Text.sln
-sudo: false
-before_install:
-  - pip install --user codecov
-after_success:
-  - codecov
