--- X.Text.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "X.Text", "X.Text.csproj", "{11D89A2E-7D6B-4E6F-B945-C8F385663658}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{11D89A2E-7D6B-4E6F-B945-C8F385663658}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{11D89A2E-7D6B-4E6F-B945-C8F385663658}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{11D89A2E-7D6B-4E6F-B945-C8F385663658}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{11D89A2E-7D6B-4E6F-B945-C8F385663658}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
