+++ src/X.Text/X.Text.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard1.3;net452</TargetFrameworks>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <Company>.NET Core Ukrainian User Group</Company>
+    <Authors>Andrew Gubskiy</Authors>
+    <Version>2.0.0</Version>
+    <PackageProjectUrl>https://github.com/ernado-x/X.Text</PackageProjectUrl>
+    <RepositoryType>git</RepositoryType>
+    <Description>Library to simplify common tasks when working with text.</Description>
+    <PackageIconUrl>https://ru.gravatar.com/userimage/8071071/f2dc08ee7e4016451f64a7ae9cffd110.png?size=200</PackageIconUrl>
+    <RepositoryUrl>https://github.com/ernado-x/X.Text.git</RepositoryUrl>
+  </PropertyGroup>
+
+</Project>
